refactor: centralize string utilities
